// src/types/meeting.ts
export interface MeetingData {
  title: string;
  description: string;
  speaker1Name: string;
  speaker1Title: string;
  speaker1Bio: string;
  speaker2Name: string;
  speaker2Title: string;
  speaker2Bio: string;
  dateTimeUS: string;
  dateTimeVN: string;
  timezone: string;
  meetingLink: string;
  bioLink: string;
  agendaLink: string;
  presentationLink: string;
  pastSessionsLink: string;
}